+++ crates/tinywasm/benches/tinywasm.rs
@@ -33,11 +33,11 @@ fn tinywasm_run(module: TinyWasmModule) -> Result<()> {
 
 fn criterion_benchmark(c: &mut Criterion) {
     let module = tinywasm_parse().expect("tinywasm_parse");
-    let twasm = tinywasm_to_twasm(&module).expect("tinywasm_to_twasm");
+    let _twasm = tinywasm_to_twasm(&module).expect("tinywasm_to_twasm");
 
-    c.bench_function("tinywasm_parse", |b| b.iter(tinywasm_parse));
-    c.bench_function("tinywasm_to_twasm", |b| b.iter(|| tinywasm_to_twasm(&module)));
-    c.bench_function("tinywasm_from_twasm", |b| b.iter(|| tinywasm_from_twasm(&twasm)));
+    // c.bench_function("tinywasm_parse", |b| b.iter(tinywasm_parse));
+    // c.bench_function("tinywasm_to_twasm", |b| b.iter(|| tinywasm_to_twasm(&module)));
+    // c.bench_function("tinywasm_from_twasm", |b| b.iter(|| tinywasm_from_twasm(&twasm)));
     c.bench_function("tinywasm", |b| b.iter(|| tinywasm_run(module.clone())));
 }