+++ crates/types/src/instructions.rs
@@ -51,7 +51,12 @@ pub enum ConstInstruction {
 #[rustfmt::skip]
 pub enum Instruction {
     LocalCopy32(LocalAddr, LocalAddr), LocalCopy64(LocalAddr, LocalAddr), LocalCopy128(LocalAddr, LocalAddr), LocalCopyRef(LocalAddr, LocalAddr),
-
+    I32AddLocals(LocalAddr, LocalAddr), I64AddLocals(LocalAddr, LocalAddr),
+    I32AddConst(i32), I64AddConst(i64),
+    I32StoreLocalLocal(MemoryArg, LocalAddr, LocalAddr),
+    I32LoadLocalTee(MemoryArg, LocalAddr, LocalAddr),
+    I64XorRotlConst(i64),
+    I64XorRotlConstTee(i64, LocalAddr),
     // > Control Instructions (jump-oriented, lowered from structured control during parsing)
     // See <https://webassembly.github.io/spec/core/binary/instructions.html#control-instructions>
     Unreachable,