+++ CHANGELOG.md
@@ -41,7 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Extern tables now correctly update their type after growing
 - Increased MSRV to 1.80.0
 - Simplify and optimize the interpreter loop
-- Use a seperate stack and locals for 32, 64 and 128 bit values and references (#21)
+- Use a separate stack and locals for 32, 64 and 128 bit values and references (#21)
 - Updated to latest `wasmparser` version
 - Removed benchmarks comparing TinyWasm to other WebAssembly runtimes to reduce build dependencies
 - Memory and Data Instances are no longer reference counted